refactor: 변경에 누락된  SelectedMember를 RoomMember로 수정 #25
