+++ app/prayers/[id]/_components/update/PrayerUpdateBody.tsx
@@ -12,15 +12,15 @@ import { usePrayerUpdateStore } from "../../../../../src/domain/prayers/stores/u
 import { useSelectedPrayerTitleStore } from "../../../../../src/domain/prayers/stores/useSelectedPrayerTitleStore";
 import { usePrayerContentsQuery } from "../../../../../src/domain/prayers/hooks/queries/usePrayerQueries";
 
-import PrayerTitleInput from "../../../_components/PrayerTitleInput";
-import PrayerMemberSelector from "../../../_components/PrayerMemberSelector";
-import PrayerContentInput from "../../../_components/PrayerContentInput";
-import PrayerAddButton from "../../../_components/PrayerAddButton";
-import PrayerCarousel from "../../../_components/PrayerCarousel";
-import PrayerDeleteDialog from "../../../_components/dialog/PrayerDeleteDialog";
-import PrayerCustomNameDialog from "../../../_components/dialog/PrayerCustomNameDialog";
-import PrayerMemberSelectionModal from "../../../_components/modal/PrayerMemberSelectionModal";
-import PrayerEditDialog from "../../../_components/dialog/PrayerEditDialog";
+import PrayerTitleInput from "../../../_components/creation/PrayerTitleInput";
+import PrayerMemberSelector from "../../../_components/creation/PrayerMemberSelector";
+import PrayerContentInput from "../../../_components/creation/PrayerContentInput";
+import PrayerAddButton from "../../../_components/creation/PrayerAddButton";
+import PrayerCarousel from "../../../_components/creation/PrayerCarousel";
+import PrayerDeleteDialog from "../../../_components/creation/dialog/PrayerDeleteDialog";
+import PrayerCustomNameDialog from "../../../_components/creation/dialog/PrayerCustomNameDialog";
+import PrayerMemberSelectionModal from "../../../_components/creation/modal/PrayerMemberSelectionModal";
+import PrayerEditDialog from "../../../_components/creation/dialog/PrayerEditDialog";
 
 const { width } = Dimensions.get("window");
 const CARD_SPACING = RFValue(20); // 카드 사이 간격
